+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/sync/model/accountdata/UserAccountData.kt
@@ -14,12 +14,19 @@
  * limitations under the License.
  */
 
-package im.vector.matrix.android.internal.session.sync.model
+package im.vector.matrix.android.internal.session.sync.model.accountdata
 
 import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 
-@JsonClass(generateAdapter = true)
-internal data class UserAccountDataSync(
-        @Json(name = "events") val list: List<UserAccountData> = emptyList()
-)
+internal abstract class UserAccountData {
+
+    @Json(name = "type") abstract val type: String
+
+    companion object {
+        const val TYPE_IGNORED_USER_LIST = "m.ignored_user_list"
+        const val TYPE_DIRECT_MESSAGES = "m.direct"
+        const val TYPE_PREVIEW_URLS = "org.matrix.preview_urls"
+        const val TYPE_WIDGETS = "m.widgets"
+        const val TYPE_PUSH_RULES = "m.push_rules"
+    }
+}
