+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/sync/model/accountdata/UserAccountDataDirectMessages.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package im.vector.matrix.android.api.util
+package im.vector.matrix.android.internal.session.sync.model.accountdata
 
-import com.squareup.moshi.Types
-import java.lang.reflect.ParameterizedType
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
 
-typealias JsonDict = Map<String, @JvmSuppressWildcards Any>
-
-val emptyJsonDict = emptyMap<String, Any>()
-
-internal val JSON_DICT_PARAMETERIZED_TYPE: ParameterizedType = Types.newParameterizedType(Map::class.java, String::class.java, Any::class.java)
+@JsonClass(generateAdapter = true)
+internal data class UserAccountDataDirectMessages(
+        @Json(name = "type") override val type: String = TYPE_DIRECT_MESSAGES,
+        @Json(name = "content") val content: Map<String, List<String>>
+) : UserAccountData()
